more style and implementation

+++ src/components/home/Snippet.js
@@ -5,8 +5,10 @@ import CodeEditor from "@uiw/react-textarea-code-editor";
 
 function Snippet({ snippet, getSnippets, editSnippet }) {
   async function deleteSnippet() {
-    await axios.delete(`http://localhost:5000/snippet/${snippet._id}`);
-    getSnippets();
+    if (window.confirm("Do you want to delete this snippet?")) {
+      await axios.delete(`http://localhost:5000/snippet/${snippet._id}`);
+      getSnippets();
+    }
   }
 
   return (
